Implement Theme flags.
Each field of a theme can have a number of flags. When a field is
queried, a value is selected based on the given set of flags.

So that combinations of flags do not have to be defined exhaustively, a
query will match the largest unambiguous subset of flags when the given
set could not be found directly.

A Theme derived from a StudioTheme selects a modifier only if exactly
one flag corresponding to a modifier is given.

Style has several new methods for setting flags on attached objects.